Moved TestDPLData into its own file from within a test class
